fleet-api/requirements.txt
fastapi
uvicorn
paho-mqtt
sqlalchemy
pydantic
sqlmodel
pytest
pytest-asyncio